apps/rag-service/requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
llama-index-core
fastmcp
llama-index-core
llama-index-llms-google-genai
llama-index-embeddings-google-genai